feat: Se crearon los repositorios para metodos personalizados, se crearon los servicios para esos metodos, los controladores estan en proceso

+++ Runeshop Ecommerce/src/main/java/com/example/runeshop_ecommerce/entities/Precio.java
@@ -14,10 +14,10 @@ import java.util.List;
 @Data
 public class Precio extends Base {
     @Column(name = "precio_compra")
-    private Float precioCompra;
+    private Double precioCompra;
 
     @Column(name = "precio_venta")
-    private Float precioVenta;
+    private Double precioVenta;
 
     @OneToMany(mappedBy = "precio")
     private List<Detalle> detalles;
